pyproject.toml
[project]
name = "marie-rag-indexing"
version = "0.1.0"
description = "A modular and scalable system for RAG indexing with multiple sources."
authors = [
    { name = "Omar Zapata", email = "andresete.chaos@gmail.com" }
]
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=6.0.0",
    "pydantic>=2.10.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "opensearch-py>=2.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "playwright>=1.40.0",
    "sqlalchemy>=2.0.30",
    "pymongo>=4.10.0",
    "boto3>=1.35.0",
    "google-api-python-client>=2.111.0",
    "google-auth-oauthlib>=1.2.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.5.0",
    "pinecone>=6.0.0",
    "qdrant-client>=1.16.2",
    "pymilvus>=2.6.6",
    "psycopg2-binary>=2.9.11",
    "pgvector>=0.4.2",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "types-requests>=2.32.4.20250913",
    "types-setuptools>=80.9.0.20251223",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["PLR0913", "PLR0915", "PLW0603"]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["__pycache__", "build", "dist", "backend/app/infrastructure/adapters/vector_stores/pgvector.py", "backend/app/infrastructure/adapters/vector_stores/pinecone.py"]
